Update bee.lua
Proposed major change to analyzeBees function, removal of ditchProduct and countBees, integrated those functions into main analyzeBees loop. Also put in remembering of what drones have been analyzed in past loops. All this minimizes time spent waiting on the analyzing process.

Please, double check this code, I have tested with many attempts and found it stable and working. I am new to LUA but not a new programmer. Also, I have another program you may want to bundle in with this project, it takes a chest full of princesses, a chest full of drones, and keeps breeding each princess to only those drones until they are an exact match for the drones (good for working pre-genetic machines). http://pastebin.com/1ufBmDXu

Also, a suggestion with this is to put the analyzer ABOVE the turtle instead of to the side, then turning of the turtle would not be needed. Another idea for when the new version of computercraft is more available, use peripheral.find to grab the analyzer.